requirements.txt
acme==1.21.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.34.1
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Automat==20.2.0
azure-core==1.30.2
azure-identity==1.17.1
Babel==2.8.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.4
build==1.2.1
cachetools==5.4.0
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
cloud-init==24.1.3
cloudpickle==3.0.0
cmake==3.30.2
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
dbus-python==1.2.18
decorator==5.1.1
demjson3==3.0.6
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==7.0.0
docstring-parser==0.15
docx2txt==0.8
exceptiongroup==1.2.1
fastapi==0.112.0
filelock==3.14.0
FLAML==2.1.2
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
googleapis-common-protos==1.56.4
gpg==1.16.0
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
hf_transfer==0.1.8
html2text==2020.1.16
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperlink==21.0.0
identify==2.5.36
idna==3.3
importlib_metadata==8.0.0
importlib_resources==6.4.2
incremental==21.3.0
interegular==0.3.3
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.3.6
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libcomps==0.1.15
litellm==1.41.12
llama-cloud==0.0.13
llama-index==0.10.65
llama-index-agent-openai==0.2.9
llama-index-cli==0.1.13
llama-index-core==0.10.66
llama-index-embeddings-azure-openai==0.1.11
llama-index-embeddings-huggingface==0.2.3
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.2.7
llama-index-legacy==0.9.48.post2
llama-index-llms-azure-openai==0.1.10
llama-index-llms-openai==0.1.29
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llvmlite==0.43.0
lm-format-enforcer==0.10.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.21.3
matplotlib==3.8.4
mdurl==0.1.2
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.40.7
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
outlines==0.0.46
overrides==7.7.0
packaging==24.0
pandas==2.2.2
parsedatetime==2.6
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.10.1
posthog==3.5.0
pre-commit==3.7.1
prettytable==3.11.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==6.0.0
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pyautogen==0.2.27
pycountry==24.6.1
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.3.0
pylance==0.8.21
pymemgpt==0.3.22
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu3
python-box==7.2.0
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
python-multipart==0.0.9
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
ratelimiter==1.2.0.post0
ray==2.34.0
referencing==0.35.1
regex==2024.4.28
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
retry==0.9.2
rich==13.7.1
rpds-py==0.19.0
rpm==4.17.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
SecretStorage==3.3.1
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
service-identity==18.1.0
sglang==0.2.13
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sos==4.5.6
soupsieve==2.5
SQLAlchemy==2.0.32
sqlalchemy-json==0.7.0
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.16
ssh-import-id==5.11
starlette==0.37.2
striprtf==0.0.26
sympy==1.13.2
systemd-python==234
tenacity==8.5.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
transformers==4.44.0
triton==3.0.0
Twisted==22.1.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.2
vllm==0.5.4
vllm-flash-attn==2.6.1
wadllib==1.3.6
WALinuxAgent==2.2.46
watchfiles==0.23.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wllvm==1.3.1
wrapt==1.16.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.9.4
zipp==1.0.0
zmq==0.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0

